+++ SentinelOne/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2025-09-19 - 1.20.15
+
+### Changed
+
+- Release of create IoCs action from Beta
+
 ## 2025-09-05 - 1.20.14
 
 ### Fixed